requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
aiohttp>=3.8.0
python-dotenv>=0.19.0
PyGithub>=1.55
